split up Client, remove Servant dep
